% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/projevan.R
\name{EAR}
\alias{EAR}
\title{Calculate the Effective Annual Rate}
\usage{
EAR(APR = 0, freq = "A")
}
\arguments{
\item{APR}{numeric. APR is the Annual Percentage Rate.}

\item{freq}{character or numeric. The frequency to calculate the Effective Annual Rate. "A"=Annual,"S"=Semi-Annual,"Q"=Quarterly,"B"=Bimonthly,"M"=Monthly,"W"=Weekly,"D"=Daily.}
}
\value{
ear numeric. The Effective Annual Rate.
}
\description{
This function calculates the Effective Annual Rate 'ear' of an annual discount/interest rate 'rate'
for a chosen frequency "freq".
}
\author{
Luis Martin Alonso
}

