% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/projevan.R
\name{fullnpv}
\alias{fullnpv}
\title{Generating a vector of Present Values from a vector of Cash Flows}
\usage{
fullnpv(x = c(-1000, 1000, 1000, 1000), rate = 0)
}
\arguments{
\item{x}{a vector of length n containing cash flows}

\item{rate}{the discount or interest rate}
}
\value{
net a vector of length n containing the present values of each component of vector 'x' at rate 'rate'
}
\description{
This function creates a vector 'net' containing the present value of the
components of vector 'x' at rate 'rate'.
}
\author{
Luis Martin Alonso
}

