% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/projevan.R
\name{payback}
\alias{payback}
\title{Calculate the Payback Period}
\usage{
payback(cf = c(-1000, 1000, 1000, 1000), rate = 0, eff = FALSE)
}
\arguments{
\item{cf}{a vector of length n containing cash flows}

\item{rate}{the discount or interest rate}

\item{eff}{logical. if eff=FALSE, payback do not count discount rate to perform its calculations.}
}
\value{
pb integer. The period when payback is completed.
}
\description{
This function calculates the Payback period 'pb' of a series of cash flows in vector
'cf' at discount/interest rate 'rate'.
}
\author{
Luis Martin Alonso
}

