% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/projevan.R
\name{plotCF}
\alias{plotCF}
\title{Plot Cash Flows}
\usage{
plotCF(cf = c(-1000, 1000, 1000, 1000), xl = "Period", yl = "Dollars")
}
\arguments{
\item{cf}{a vector of length n containing cash flows}

\item{xl}{X label}

\item{yl}{Y label}
}
\value{
a barplot of the Cash Flows
}
\description{
This function plots the Cash Flows in their corresponding periods.
}
\author{
Luis Martin Alonso
}
\seealso{
\code{barplot}
}

