% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/projevan.R
\name{plotRate}
\alias{plotRate}
\title{Plot the NPV profile}
\usage{
plotRate(cf = c(-1000, 1000, 1000, 1000))
}
\arguments{
\item{cf}{a vector of length n containing cash flows}
}
\value{
a plot of NPV vs. Interest Rate.
}
\description{
This function plot the Net Present Value (NPV) of a series of cash flows in vector 'cf' to a series of
interest/discount rates.
}
\author{
Luis Martin Alonso
}
\seealso{
\code{plot,abline}
}

