% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/projevan.R
\name{irr}
\alias{irr}
\title{Calculate the Internal Rate of Return}
\usage{
irr(x = c(-1000, 1000, 1000, 1000))
}
\arguments{
\item{x}{a vector of length n containing cash flows}
}
\value{
rate a numeric value representing the Internal Rate of Return (IRR) of Cash Flows in 'x'.
}
\description{
This function calculates the Internal Rate of Return (IRR) 'rate' of a series of cash flows in vector
'x'.
}
\author{
Luis Martin Alonso
}

