% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/projevan.R
\name{npv}
\alias{npv}
\title{Calculate the Net Present Value}
\usage{
npv(cf = c(-1000, 1000, 1000, 1000), rate = 0)
}
\arguments{
\item{cf}{a vector of length n containing cash flows}

\item{rate}{the discount or interest rate}
}
\value{
net a numeric value representing the Net Present Value of Cash Flows in 'cf'
}
\description{
This function calculates the Net Present Value (NPV) 'net' of a series of cash flows in vector
'cf' at discount/interest rate 'rate'.
}
\author{
Luis Martin Alonso
}

